[refactor] cache package
